Fixing dead (404) `manage-git-clones` link
The `README.md` previously referenced a link to the `manage-git-clones` repo,
which no longer exists.

I am submitting a patch to change the link to the `manage-git-branches` repo,
which I believe is the successor to `manage-git-clones`. Please let me know if
this is not the case so I can update this PR with the correct repository link.